+++ tests/testthat/test_expncat_stat.R
@@ -12,7 +12,7 @@ test_that("test output object", {
   X <- mtcars$gear
   Y <- mtcars$am
   catstat <- ExpStat(X, Y, valueOfGood = 1)
-  ctv <- chisq.test(X, Y)
+  ctv <- chisq.test(X, Y, simulate.p.value = TRUE)
   ctat <- as.numeric(round(ctv$statistic, 3))
   expect_output(str(catstat), "chr")
   expect_equal(ctat, as.numeric(catstat[2]))